api/studentapi/requirements.txt
fastapi 
uvicorn 
sqlmodel 
asyncpg
psycopg2-binary
pyyaml
python-dotenv
openai 

